requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
adafruit-blinka==8.56.0
    # via
    #   adafruit-circuitpython-busdevice
    #   adafruit-circuitpython-connectionmanager
    #   adafruit-circuitpython-requests
    #   adafruit-circuitpython-scd4x
    #   adafruit-circuitpython-typing
adafruit-circuitpython-busdevice==5.2.11
    # via
    #   adafruit-circuitpython-scd4x
    #   adafruit-circuitpython-typing
adafruit-circuitpython-connectionmanager==3.1.3
    # via adafruit-circuitpython-requests
adafruit-circuitpython-requests==4.1.10
    # via adafruit-circuitpython-typing
adafruit-circuitpython-scd4x==1.4.6
    # via -r requirements.in
adafruit-circuitpython-typing==1.11.2
    # via
    #   adafruit-blinka
    #   adafruit-circuitpython-busdevice
adafruit-platformdetect==3.77.0
    # via adafruit-blinka
adafruit-pureio==1.1.11
    # via adafruit-blinka
bidict==0.23.1
    # via python-socketio
binho-host-adapter==0.1.6
    # via adafruit-blinka
blinker==1.9.0
    # via flask
board==1.0
    # via -r requirements.in
click==8.1.8
    # via flask
colorama==0.4.6
    # via click
colorzero==2.0
    # via gpiozero
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-socketio
flask-socketio==5.5.1
    # via -r requirements.in
gpiozero==2.0.1
    # via -r requirements.in
h11==0.14.0
    # via wsproto
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
pigpio==1.78
    # via -r requirements.in
pyftdi==0.56.0
    # via adafruit-blinka
pyserial==3.5
    # via
    #   binho-host-adapter
    #   pyftdi
python-engineio==4.11.2
    # via python-socketio
python-socketio==5.12.1
    # via flask-socketio
pyusb==1.3.1
    # via pyftdi
rpi-gpio==0.7.1
    # via -r requirements.in
rpi-hardware-pwm==0.2.2
    # via -r requirements.in
simple-websocket==1.1.0
    # via python-engineio
typing-extensions==4.13.1
    # via adafruit-circuitpython-typing
werkzeug==3.1.3
    # via flask
wsproto==1.2.0
    # via simple-websocket

# The following packages are considered to be unsafe in a requirements file:
# setuptools
